renames MenuItem to MenuDropdown
