+++ components/src/NavBar/DesktopMenu.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
 import { Box } from "ComponentsRoot";
-import { MenuItem } from "./MenuItem";
+import { MenuDropdown } from "./MenuDropdown";
 import SubMenuItem from "./SubMenuItem";
 import theme from "../theme";
 
@@ -11,14 +11,14 @@ const BaseDesktopMenu = ({
   ...props
 }) => (
   <Box { ...props }>
-    {Object.entries(menuData).map(([menuItemKey, menuItem]) => (
-      <MenuItem key={ menuItemKey } labelText={ menuItemKey }>
-        {Object.entries(menuItem).map(([subMenuItemKey, subMenuItem]) => (
+    {Object.entries(menuData).map(([menuDropdownKey, menuDropdown]) => (
+      <MenuDropdown key={ menuDropdownKey } labelText={ menuDropdownKey }>
+        {Object.entries(menuDropdown).map(([subMenuItemKey, subMenuItem]) => (
           <SubMenuItem key={ subMenuItemKey } href={ subMenuItem.href } subText={ subMenuItem.subText }>
             { subMenuItemKey }
           </SubMenuItem>
         ))}
-      </MenuItem>
+      </MenuDropdown>
     ))}
   </Box>
 );
