+++ components/src/NavBar/MenuDropdown/index.js
@@ -1,3 +1,3 @@
-export { default as MenuItem } from "./MenuItem";
+export { default as MenuDropdown } from "./MenuDropdown";
 export { default as SubMenu } from "./SubMenu";
 export { default as SubMenuItemList } from "./SubMenuItemList";
